# docs/api/openapi/plan-service.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "혼디가개 여행 일정 서비스 API 명세서",
    "description": "Plan Service 전용",
    "version": "v1"
  },
  "servers": [
    {
      "url": "/",
      "description": "Plan Service"
    }
  ],
...
  "paths": {
...
    "/api/v1/plans/{planId}/walk-safety": {
      "get": {
        "tags": [
          "일정 산책 위험도"
        ],
        "summary": "일정 항목 산책 위험도",
        "description": "일정 항목마다 **그 항목의 시각 기준으로** 산책 위험도를 판정합니다. 일자 날씨 브리핑(`/weather`)이 \"둘째 날 괜찮아?\" 에 답한다면 이 API 는 \"두 시에 그 해수욕장 걸어도 돼?\" 에 답합니다 — 산책 위험도는 시각에 따라 갈려 일자로 접을 수 없습니다. 판정 규칙(노면온도 추정·체감온도·등급 임계)은 장소 산책 위험도 API 와 **같은 것**을 씁니다. 판정 기준 반려견(basisPetId)은 **그날 날씨 판정의 기준과 같습니다** — 두 화면이 같은 날을 다른 아이 기준으로 말하지 않게 합니다.\n\n판정을 못 낸 항목도 줄은 그대로 내려가고 unavailableReasonCode 와 unavailableReason(문장)이 짝으로 옵니다. 코드는 다섯으로 갈립니다 — PAST_DATE(지난 날짜, 다시 물어도 생기지 않습니다) · NOT_PLACE_TARGET(좌표를 아는 장소 항목이 아님, 예: 올레 코스·이동) · NO_START_TIME(항목에 시각이 없음 — **없는 시각을 정오 등으로 지어내지 않습니다**) · BEYOND_FORECAST_RANGE(시각별 예보 범위 밖, 기다리면 풀립니다) · LOOKUP_FAILED(조회 실패, 다섯 중 이것만 일시적 장애입니다). 재시도 안내는 이 코드로 판단하고 문장을 파싱하지 않습니다.\n\n**시각별 예보가 닿는 범위는 `오늘 ~ 오늘+4`(5일)로, 일자 날씨 브리핑의 11일보다 짧습니다.** 산책 위험도는 노면온도를 시각·일사로 추정해야 해서 시각별 데이터가 있는 단기예보만 쓰고, 중기예보(오늘+4~오늘+10)에는 오전/오후뿐이라 오후 두 시 아스팔트를 계산할 수 없기 때문입니다. 그 밖의 날짜는 원격 호출 없이 BEYOND_FORECAST_RANGE 로 내려갑니다 — 같은 일정이라도 `/weather` 에는 판정이 있는데 여기는 비어 있는 날이 **정상**입니다.\n\n근거 목록·시간대별 판정·특보 상세는 이 응답에 없습니다 — 항목마다 실어 나르면 응답이 항목 수만큼 부풀기 때문입니다. 필요하면 장소 산책 위험도 API(`GET /api/v1/places/{placeId}/walk-safety`)를 직접 부르세요. 판정을 못 낸 줄에도 placeId 는 남습니다(NOT_PLACE_TARGET 제외) — 그래야 그 줄에서 직접 부를 수 있습니다.\n\n**필수: planId (경로).**\n\n호출 예\n- `GET /api/v1/plans/1234567890123456789/walk-safety`",
        "operationId": "getWalkSafety",
        "parameters": [
          {
            "name": "planId",
            "in": "path",
            "description": "[필수] 일정 아이디. Snowflake 숫자라 환경(dev/prod)마다 다르고 예시 값은 형식 안내용입니다. 실제 값은 일정 목록 응답의 planId 를 그대로 씁니다",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1234567890123456800
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponsePlanWalkSafetyResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataHeader": {
        "type": "object",
        "description": "공통 응답 헤더",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "성공 여부"
          },
          "resultCode": {
            "type": "string",
            "description": "대표 오류 코드. 성공이면 null"
          },
          "resultMessage": {
            "type": "string",
            "description": "사용자에게 보여줄 대표 오류 메시지. 성공이면 null. 오류 종류와 무관하게 항상 문자열이다"
          },
          "fieldErrors": {
            "type": "array",
            "description": "필드 단위 검증 오류 목록. 검증 실패가 아니면 null",
            "items": {
              "$ref": "#/components/schemas/ValidationErrorItem"
            }
          }
        }
      },
...
      "ValidationErrorItem": {
        "type": "object",
        "description": "필드 단위 검증 오류 항목",
        "properties": {
          "code": {
            "type": "string",
            "description": "오류 코드 (도메인별 검증 코드)",
            "example": "MEMBER_109"
          },
          "field": {
            "type": "string",
            "description": "오류가 발생한 필드명",
            "example": "nickname"
          },
          "message": {
            "type": "string",
            "description": "오류 메시지",
            "example": "닉네임은 10자 이하만 가능합니다."
          }
        }
      },
...
      "ScoreMetricMetadata": {
        "type": "object",
        "description": "점수 지표 메타데이터 DTO",
        "properties": {
          "code": {
            "type": "string",
            "description": "내부 코드 값",
            "example": "OPPORTUNITY_SCORE"
          },
          "name": {
            "type": "string",
            "description": "화면 표시 이름",
            "example": "기회도"
          },
          "description": {
            "type": "string",
            "description": "지표 설명"
          },
          "scoreDescription": {
            "type": "string",
            "description": "점수 해석 설명"
          }
        }
      },
      "PlanItemWalkSafetyItem": {
        "type": "object",
        "description": "일정 항목 산책 위험도 DTO",
        "properties": {
          "planItemId": {
            "type": "string",
            "description": "일정 항목 아이디",
            "example": 1234567890123456800
          },
          "day": {
            "type": "integer",
            "format": "int32",
            "description": "일차 (1부터)",
            "example": 2
          },
          "date": {
            "type": "string",
            "format": "date",
            "description": "그 일차의 날짜",
            "example": "2026-09-13"
          },
          "sequence": {
            "type": "integer",
            "format": "int32",
            "description": "그 일차 안의 순서 (0부터)",
            "example": 1
          },
          "startTime": {
            "type": "string",
            "description": "항목 시작 시각. 없으면 null 이고 위험도도 비어 있다",
            "example": "14:00:00"
          },
          "title": {
            "type": "string",
            "description": "항목 이름",
            "example": "협재해수욕장"
          },
          "placeId": {
            "type": "string",
            "description": "장소 아이디. 좌표를 아는 장소 항목이 아니면(unavailableReasonCode=NOT_PLACE_TARGET) null 입니다. **판정을 못 낸 줄에도 남습니다** — LOOKUP_FAILED 같은 줄에서 화면이 `GET /api/v1/places/{placeId}/walk-safety` 를 직접 불러 다시 시도할 수 있어야 하기 때문입니다",
            "example": 212481712381923330
          },
          "placeTitle": {
            "type": "string",
            "description": "tour-service 가 확인한 장소명. **판정을 못 낸 항목은 물어보지 않았으므로 null 입니다** — 일정에 적힌 이름은 title 로 내려갑니다",
            "example": "협재해수욕장"
          },
          "targetDateTime": {
            "type": "string",
            "format": "date-time",
            "description": "판정 기준 시각",
            "example": "2026-09-13T14:00:00"
          },
          "basisPetId": {
            "type": "string",
            "description": "판정 기준 반려견 아이디. 그날 날씨 판정의 basisPetId 와 같다",
            "example": 1234567890123456800
          },
          "walkSafetyLevel": {
            "$ref": "#/components/schemas/ScoreMetricMetadata",
            "description": "산책 위험도 등급 metadata. 못 낸 항목은 null"
          },
          "estimatedPavementCelsius": {
            "type": "number",
            "format": "double",
            "description": "추정 노면(아스팔트) 표면온도(섭씨). 실측이 아니라 tour-service 의 추정치다. **화면 문구에 '노면(아스팔트)' 를 밝히고 temperature(기온)와 나란히 보여 주세요** — 노면온도만 표시하면 사용자가 그것을 기온으로 읽는다",
            "example": 58
          },
          "feelsLikeCelsius": {
            "type": "number",
            "format": "double",
            "description": "기상청 여름철 체감온도(섭씨). 등급 판정의 기준값이다",
            "example": 33.5
          },
          "temperature": {
            "type": "number",
            "format": "double",
            "description": "판정에 쓴 기온(섭씨)",
            "example": 31
          },
          "saferWindowStart": {
            "type": "string",
            "description": "같은 날 더 안전한 시간대 시작. 없으면 null",
            "example": "18:00:00"
          },
          "saferWindowEnd": {
            "type": "string",
            "description": "같은 날 더 안전한 시간대 종료. 없으면 null",
            "example": "21:00:00"
          },
          "unavailableReasonCode": {
            "type": "string",
            "description": "판정을 못 낸 사유 코드. 정상이면 null. PAST_DATE 지난 날짜 · NOT_PLACE_TARGET 장소 항목 아님 · NO_START_TIME 시각 미지정 · BEYOND_FORECAST_RANGE 시각별 예보 범위(오늘~오늘+4) 밖 · LOOKUP_FAILED 조회 실패",
            "example": "NO_START_TIME"
          },
          "unavailableReason": {
            "type": "string",
            "description": "판정을 못 낸 사유 문장. 정상이면 null",
            "example": "이 항목에 시작 시각이 없어 산책 위험도를 낼 수 없습니다. 시각은 시간대마다 판정이 갈립니다."
          }
        }
      },
      "PlanWalkSafetyResponse": {
        "type": "object",
        "description": "일정 항목 산책 위험도 응답 DTO",
        "properties": {
          "planId": {
            "type": "string",
            "description": "일정 아이디",
            "example": 1234567890123456800
          },
          "planTitle": {
            "type": "string",
            "description": "일정 제목",
            "example": "몽실이와 제주 2박 3일"
          },
          "petIds": {
            "type": "array",
            "description": "판정에 들어간 동행 반려견 아이디 목록. 항목별 basisPetId 는 이 안의 하나다",
            "example": [
              "1234567890123456789"
            ],
            "items": {
              "type": "string"
            }
          },
          "items": {
            "type": "array",
            "description": "일차·순서대로 정렬된 항목별 판정. 못 낸 항목도 사유와 함께 그대로 들어간다",
            "items": {
              "$ref": "#/components/schemas/PlanItemWalkSafetyItem"
            }
          }
        }
      },
      "ResponsePlanWalkSafetyResponse": {
        "type": "object",
        "description": "공통 응답 래퍼",
        "properties": {
          "dataHeader": {
            "$ref": "#/components/schemas/DataHeader",
            "description": "응답 헤더"
          },
          "dataBody": {
            "$ref": "#/components/schemas/PlanWalkSafetyResponse",
            "description": "응답 본문. 실패면 null"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}